+++ src/image.ts
@@ -2,6 +2,9 @@
 
 export {
   Image,
+  ImageContent,
+  ImageFallback,
+  ImageRoot,
   type ImageContentProps,
   type ImageFallbackProps,
   type ImageFit,